'CMakeLists.txt' (TrkExStraightLineIntersector-01-00-03)
	* TrkExStraightLineIntersector/StraightLineIntersector.h: 64-bit counters (except serialNumber)
	* tag as TrkExStraightLineIntersector-01-00-03

+++ Tracking/TrkExtrapolation/TrkExStraightLineIntersector/TrkExStraightLineIntersector/StraightLineIntersector.h
@@ -85,7 +85,7 @@ private:
     double			m_transverseLength;
 
     // counters
-    int				m_countExtrapolations;
+    unsigned long long		m_countExtrapolations;
     
 };
